xero_python/requirements.txt
requests-oauthlib
requests
pydantic
python-dateutil
urllib3
PyJWT

google.auth
google-cloud-secret-manager
google-cloud-storage
ratelimit